--- src/test/java/com/example/springbootboard/controller/UserControllerTest.java
@@ -1,4 +1,0 @@
-package com.example.springbootboard.controller;
-
-public class UserControllerTest {
-}
