+++ src/test/java/com/example/springbootboard/domain/users/UserControllerTest.java
@@ -1,0 +1,15 @@
+package com.example.springbootboard.domain.users;
+
+import org.junit.jupiter.api.Test;
+
+class UserControllerTest {
+    @Test
+    void name() {
+        // given
+
+        // when
+
+        // then
+    }
+
+}
